+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/gui/Window.h
@@ -868,7 +868,7 @@ namespace gui
       static_cast<CWindow*>(cbhdl)->GetContextButtons(itemNumber, buttonList);
       if (!buttonList.empty())
       {
-        unsigned int presentSize = buttonList.size();
+        unsigned int presentSize = static_cast<unsigned int>(buttonList.size());
         if (presentSize > *size)
           kodi::Log(ADDON_LOG_WARNING, "GetContextButtons: More as allowed '%i' entries present!", *size);
         else
